Update to work with Billboard's new UI
Billboard has a flashy new UI, and that change broke this module.

This is the beginning of a fix for it: as of right now, this has basic function for everything EXCEPT the artist charts. I will look into that tomorrow. 

Would it be possible to merge this fix and deploy to PyPI so that there is at least limited functionality while I complete the fix? This is blocking me on another project.