Trying out subclasses
Just learned about subclasses so trying it out. previousDate and nextDate work now but it just kind of keeps going since currently there is no solid way to stop it. Thinking about doing what I did if someone enter a date that is too far in the past or future. Normal billboard charts redirect to the newest or oldest possible date if you enter one too old or too new but year end charts don't do this so added some stuff to help with this.