Add the title attribute to ChartData
This commit adds a "title" attribute to instances of the ChartData class. This
attribute contains the human-readable title of the chart,
e.g. "The Hot 100" or "Digital Albums".

There is no one place in the chart pages where the chart title is
consistently given across all charts, with proper casing. Therefore we
look at both the alt text of the h1 image, if it exists, and otherwise
we take the all-caps text from the h1 element and do our own title
casing on it.

The title casing logic does not handle all possible strings, but it
works with nearly all known Billboard chart names. The only exceptions
are the charts with "LyricFind" in the name, which will be rendered
"Lyricfind".