Use the 'twitter:title' meta tag for chart titles
Instead of using the name element in the page body, which often has the
chart title in all caps and requires us to do the title casing
ourselves, we can use the `twitter:title` meta tag. The only catch to
this is at least one chart name is formatted like this:
`The Hot 100 | Billboard`. Since we only want the chart title, we strip
off the `|` character and everything after it.