Handle Charts without an "award" column
As best I can tell, this really only applies to the current TikTok chart. 

The issue is that the column numbers are off because of the the Nth column of metadata isn't found. 

I handled this by detecting the presence of the award column, and if missing adjusting the column indices down. 

* Haven't added tests.
* Need to adapt this to year end charts, in theory. 

A more stable change would be for all metadata indexing to determine the relative column position but that's probably not necessary.